+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Api.cs
@@ -9,6 +9,7 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 using Piipan.Match.Func.Api.DataTypeHandlers;
+using Piipan.Match.Func.Api.Models;
 using Piipan.Match.Func.Api.Parsers;
 using Piipan.Match.Func.Api.Resolvers;
 using Piipan.Match.Shared;
