+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Models/OrchMatchResult.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using Piipan.Participants.Api.Models;
 using Newtonsoft.Json;
 
-namespace Piipan.Match.Func.Api
+namespace Piipan.Match.Func.Api.Models
 {
 
     /// <summary>
