+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Resolvers/IMatchResolver.cs
@@ -1,7 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
+using Piipan.Match.Func.Api.Models;
 
 namespace Piipan.Match.Func.Api.Resolvers
 {
