+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Validators/RequestPersonPiiValidator.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Piipan.Match.Func.Api.Models;
 using FluentValidation;
 
 namespace Piipan.Match.Func.Api.Validators
