+++ match/tests/Piipan.Match.Func.Api.Tests/Validators/RequestPersonPiiValidatorTests.cs
@@ -1,7 +1,6 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
+using Piipan.Match.Func.Api.Models;
 using FluentValidation.TestHelper;
 using Xunit;
