+++ match/tests/Piipan.Match.Func.Api.Tests/Validators/RequestPersonValidatorTests.cs
@@ -1,3 +1,4 @@
+using Piipan.Match.Func.Api.Models;
 using FluentValidation.TestHelper;
 using Xunit;
